Добавляет EventController и подключает его к TripController
